add required check workflow
this is the first part of a change to make tacos apply
depend on required checks passing

it adds a new `required-check` workflow. once this is merged,
it also needs to be marked as "required" in github.

this is needed in order to be able to write a test case that
will fail to perform tacos apply when this check fails.