fix: gcloud-auth-export-access-token under fish shell (#265)
was provisioning a terraform sandbox and ran into an unbound var error
for CLOUDSDK_AUTH_ACCESS_TOKEN

this is because gcloud-auth-export-access-token outputs correct fish
syntax which is then passed to an eval in sudo-gcp, however sudo-gcp is
always running under `/bin/bash`. So we should just always give it
correct bash instead.