Allow anyone to use the :taco::apply label, not just the PR owner.
This was originally to have the PR owner take resposibility for
approving. The intention was that if another user wanted to apply, they
could close the PR, leaving the branch open, and the open a new PR on
that branch which they owned. In practice, however, this proved to be
too much friction.

Fixes #259